Fixes TS errors for NodeModulesMode
