+++ packages/plugin-nm/sources/core.d.ts
@@ -1,0 +1,9 @@
+import '@yarnpkg/core';
+
+declare module '@yarnpkg/core' {
+  interface ConfigurationValueMap {
+    nmHoistingLimits: NodeModulesHoistingLimits;
+    nmMode: NodeModulesMode;
+    nmSelfReferences: boolean;
+  }
+}
